Refactor: Create core-datasource module
This commit creates a new `core-datasource` module and moves relevant files into it.

The following files and classes were moved:
- `ItineraryItem.kt`
- `InsertionSuggestion.kt`
- `DatasourceModule.kt`
- `Prompts.kt`
- `VertexProvider.kt` and `VertexProviderImpl.kt`

The app module now depends on the `core-datasource` module.
The project name has been updated in `.idea/.name`.
Appropriate Gradle configurations and dependencies have been added for the new module.
