+++ .idea/.name
@@ -1,1 +1,1 @@
-Itinerar-AI
+What beats rock
