+++ build.gradle.kts
@@ -6,4 +6,5 @@ plugins {
     alias(libs.plugins.ksp) apply false
     alias(libs.plugins.google.gms.google.services) apply false
     alias(libs.plugins.hilt.android) apply false
+    alias(libs.plugins.android.library) apply false
 }
