+++ core-datasource/src/main/java/ke/don/core_datasource/di/DatasourceModule.kt
@@ -1,9 +1,11 @@
-package ke.don.what_beats_rock.di
+package ke.don.core_datasource.di
 
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import ke.don.core_datasource.ai.VertexProvider
+import ke.don.core_datasource.ai.VertexProviderImpl
 import javax.inject.Singleton
 
 @Module
