+++ gradle/libs.versions.toml
@@ -20,6 +20,8 @@ kotlinxSerializationJson = "1.6.3"
 firebaseAppcheckKtx = "18.0.0"
 firebaseAppcheckPlayintegrity = "18.0.0"
 firebaseAppcheckDebug = "18.0.0"
+appcompat = "1.7.1"
+material = "1.12.0"
 
 
 [libraries]
@@ -49,6 +51,8 @@ kotlinx-serialization-json = { module = "org.jetbrains.kotlinx:kotlinx-serializa
 firebase-appcheck-ktx = { group = "com.google.firebase", name = "firebase-appcheck-ktx", version.ref = "firebaseAppcheckKtx" }
 firebase-appcheck-playintegrity = { group = "com.google.firebase", name = "firebase-appcheck-playintegrity", version.ref = "firebaseAppcheckPlayintegrity" }
 firebase-appcheck-debug = { group = "com.google.firebase", name = "firebase-appcheck-debug", version.ref = "firebaseAppcheckDebug" }
+androidx-appcompat = { group = "androidx.appcompat", name = "appcompat", version.ref = "appcompat" }
+material = { group = "com.google.android.material", name = "material", version.ref = "material" }
 
 [bundles]
 hilt = [
@@ -66,5 +70,6 @@ google-gms-google-services = { id = "com.google.gms.google-services", version.re
 hilt-android= {id = "com.google.dagger.hilt.android", version.ref ="hiltAndroid"}
 kotlin-serialization = { id = "org.jetbrains.kotlin.plugin.serialization", version.ref = "kotlin" }
 ksp = {id = "com.google.devtools.ksp", version.ref = "ksp"}
+android-library = { id = "com.android.library", version.ref = "agp" }
